requirements.txt
ethereum-etl
google-cloud-bigquery
pandas